requirements.txt
amqp==5.1.1
aniso8601==9.0.1
arrow==1.3.0
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.7.0
dateparser==1.2.0
dnspython==2.6.1
email_validator==2.1.1
Flask==2.3.3
Flask-Caching==2.0.2
Flask-Celery==2.4.3
Flask-Cors==4.0.0
Flask-Excel==0.0.7
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.3
Flask-Principal==0.4.0
flask-redis==0.4.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
Flask-WTF==1.2.1
flask_weasyprint==1.1.0
flower==2.0.1
fonttools==4.46.0
greenlet==2.0.2
html5lib==1.1
humanize==4.8.0
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
kombu==5.3.2
lml==0.1.0
logger==1.4
lorem==0.1.1
MarkupSafe==2.1.3
moment==0.12.1
passlib==1.7.4
Pillow==10.1.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
pydyf==0.8.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
texttable==1.7.0
times==0.7
tinycss2==1.2.1
tornado==6.3.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.1
vine==5.0.0
wcwidth==0.2.8
weasyprint==60.2
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.1.2
zopfli==0.2.3
